🗂️ types(none): improve typechecking
