+++ sources/@roots/bud-babel/package.json
@@ -48,22 +48,16 @@
   ],
   "type": "module",
   "exports": {
-    ".": {
-      "import": "./lib/index.js",
-      "default": "./lib/index.js"
-    },
-    "./types": {
-      "import": "./lib/types.js",
-      "default": "./lib/types.js"
-    }
+    ".": "./lib/index.js",
+    "./extension": "./lib/extension.js"
   },
   "typesVersions": {
     "*": {
       ".": [
         "./lib/index.d.ts"
       ],
-      "types": [
-        "./lib/types.d.ts"
+      "extension": [
+        "./lib/extension.d.ts"
       ]
     }
   },
